Add TypeScript screen validation pipeline runner
