+++ acumate-plugin/readme.md
@@ -135,5 +135,5 @@ The **AcuMate** extension provides several commands to streamline development ta
    - Inside VS Code, run **AcuMate: Validate Screens (HTML)** to queue the validator against all HTML files beneath `src/screens` (or any folder you input). A cancellable progress notification tracks the run, and results are aggregated in the **AcuMate Validation** output channel so you can inspect warnings without breaking your workflow.
    - For TypeScript coverage, run **AcuMate: Validate Screens (TypeScript)** to traverse the same folder structure, execute `collectGraphInfoDiagnostics` for each screen `.ts`, and summarize backend metadata mismatches in the output channel (requires `acuMate.useBackend = true`). This command is also cancellable so you can stop long-running validations instantly.
    - From a pipeline, run `npm run validate:screens -- <screenRoot> <workspaceRoot>` to execute the HTML scan inside the VS Code Extension Host without running the whole extension test suite. You can also set `SCREEN_VALIDATION_ROOT`, `SCREEN_VALIDATION_WORKSPACE_ROOT`, and `VSCODE_EXECUTABLE_PATH` when arguments are less convenient. Set `SCREEN_VALIDATION_FAIL_ON_DIAGNOSTICS=true` or pass `--fail-on-diagnostics` when diagnostics should fail the step.
-   - For TypeScript coverage, run `npm run validate:screens:ts` to execute the backend-powered scan through the VS Code test runner. Override the root with `TS_SCREEN_VALIDATION_ROOT` when needed. The script logs summaries instead of failing on warnings.
+   - For TypeScript coverage, run `npm run validate:screens:ts -- <screenRoot> <workspaceRoot>` to execute the backend-powered scan inside the VS Code Extension Host without running the whole extension test suite. You can also set `TS_SCREEN_VALIDATION_ROOT`, `TS_SCREEN_VALIDATION_WORKSPACE_ROOT`, and `VSCODE_EXECUTABLE_PATH`; set `TS_SCREEN_VALIDATION_FAIL_ON_DIAGNOSTICS=true` or pass `--fail-on-diagnostics` when diagnostics should fail the step.
